+++ src/lib/services/username/index.ts
@@ -5,17 +5,17 @@ import {
   useInitia,
   useValidateAddress,
 } from "lib/app-provider";
+import { useL1InfoByNetworkType } from "lib/hooks";
 import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
 import type { Addr, Nullish } from "lib/types";
 
-import { useL1InfoByNetworkType } from "./hooks";
 import { getAddressByInitiaUsername, getInitiaUsernameByAddress } from "./rest";
 
 export const useInitiaUsernameByAddress = (
   address: Nullish<Addr>,
   enabled = true
 ) => {
-  const getL1InfoByNetworkType = useL1InfoByNetworkType();
+  const { l1Rest, l1Usernames } = useL1InfoByNetworkType();
   const isInitia = useInitia();
   const { isSomeValidAddress } = useValidateAddress();
   const formatAddress = useFormatAddresses();
@@ -24,7 +24,6 @@ export const useInitiaUsernameByAddress = (
     if (!address)
       throw new Error("address is undefined (useInitiaUsernameByAddress)");
 
-    const { l1Rest, l1Usernames } = getL1InfoByNetworkType();
     const username = await getInitiaUsernameByAddress(
       l1Rest,
       l1Usernames,
@@ -48,12 +47,11 @@ export const useAddressByInitiaUsername = (
   username: string,
   enabled = true
 ) => {
-  const getL1InfoByNetworkType = useL1InfoByNetworkType();
+  const { l1Rest, l1Usernames } = useL1InfoByNetworkType();
   const isInitia = useInitia();
   const formatAddress = useFormatAddresses();
 
   const queryFn = async () => {
-    const { l1Rest, l1Usernames } = getL1InfoByNetworkType();
     const address = await getAddressByInitiaUsername(
       l1Rest,
       l1Usernames,
